feat/GameFlow : Game over loop
Game Over retry buttons now correctly starts a new game
